third_party/blink/web_tests/platform/generic/http/tests/websocket/url-with-empty-query-expected.txt

Make sure handshake with URL with empty query components success.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

url=ws://127.0.0.1:8880/echo-path?
WebSocket is open
received:/echo-path?
WebSocket is closed
PASS handshake_success is true
PASS path is expected_path
PASS successfullyParsed is true

TEST COMPLETE
